connectly_ztry/connectly_mongo/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
flask-pymongo==2.3.0
pymongo==4.7.3
python-socketio==5.11.3
python-engineio==4.9.1
eventlet==0.35.2
Pillow==10.3.0
python-dotenv==1.0.1
groq>=0.9.0
google-genai>=1.0.0
openai==1.30.1
gunicorn==22.0.0
Werkzeug==3.0.3
dnspython==2.6.1
